refactor(oauth) : 리팩토링 - 의미 있는 메서드 명으로 변경 - 변수 인라인화 - 파일 위치 이동
