+++ src/main/java/com/example/minicommerce/client/NaverOauthClient.java
@@ -8,7 +8,7 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClient;
 
-import com.example.minicommerce.OauthLoginResponseDto;
+import com.example.minicommerce.global.GetOauthTokenResponseDto;
 import com.example.minicommerce.global.ProviderType;
 
 @Component
@@ -39,7 +39,7 @@ public class NaverOauthClient implements OauthClient {
       - state: 콜백으로 받은 state
     */
     @Override
-    public OauthLoginResponseDto login(String code, String state) {
+    public GetOauthTokenResponseDto getToken(String code, String state) {
         MultiValueMap<String, String> form = new LinkedMultiValueMap<>();
         form.add("grant_type", "authorization_code");
         form.add("client_id", naverClientId);
@@ -47,13 +47,11 @@ public class NaverOauthClient implements OauthClient {
         form.add("code", code);
         form.add("state", state);
 
-        OauthLoginResponseDto body = restClient.post()
+        return restClient.post()
             .uri(naverTokenUri)
             .contentType(APPLICATION_FORM_URLENCODED)
             .body(form)
             .retrieve()
-            .body(OauthLoginResponseDto.class);
-
-        return body;
+            .body(GetOauthTokenResponseDto.class);
     }
 }
