+++ src/main/java/com/example/minicommerce/client/OauthClient.java
@@ -1,9 +1,9 @@
 package com.example.minicommerce.client;
 
-import com.example.minicommerce.OauthLoginResponseDto;
+import com.example.minicommerce.global.GetOauthTokenResponseDto;
 import com.example.minicommerce.global.ProviderType;
 
 public interface OauthClient {
     ProviderType provider();
-    OauthLoginResponseDto login(String code, String state);
+    GetOauthTokenResponseDto getToken(String code, String state);
 }
