+++ src/main/java/com/example/minicommerce/controller/OauthController.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.minicommerce.OauthLoginResponseDto;
+import com.example.minicommerce.global.GetOauthTokenResponseDto;
 import com.example.minicommerce.service.OauthService;
 
 @RestController
@@ -18,11 +18,11 @@ public class OauthController {
     }
 
     @GetMapping("/{provider}")
-    public OauthLoginResponseDto oauthLogin(
+    public GetOauthTokenResponseDto getOauthToken(
         @PathVariable String provider,
         @RequestParam(required = false) String code,
         @RequestParam(required = false) String state
     ) {
-        return oauthService.oauthLogin(provider, code, state);
+        return oauthService.getOauthToken(provider, code, state);
     }
 }
