+++ src/main/java/com/example/minicommerce/service/OauthService.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import org.springframework.stereotype.Service;
 
-import com.example.minicommerce.OauthLoginResponseDto;
+import com.example.minicommerce.global.GetOauthTokenResponseDto;
 import com.example.minicommerce.global.ProviderType;
 import com.example.minicommerce.client.OauthClient;
 
@@ -22,8 +22,8 @@ public class OauthService {
         }
     }
 
-    public OauthLoginResponseDto oauthLogin(String provider, String code, String state) {
+    public GetOauthTokenResponseDto getOauthToken(String provider, String code, String state) {
         OauthClient oauthClient = clients.get(ProviderType.valueOf(provider.toUpperCase()));
-        return oauthClient.login(code, state);
+        return oauthClient.getToken(code, state);
     }
 }
